currentStreamId type uint256 was changed to uint64,and the streamID type was changed to uint64

+++ contracts/Streaming.sol
@@ -5,9 +5,9 @@ contract Streaming {
     uint64 public streamIdCounter;
     address public owner;
 
-    mapping(uint256 => Stream) private streams;
+    mapping(uint64 => Stream) private streams;
 
-    modifier onlySenderOrRecipient(uint256 streamId) {
+    modifier onlySenderOrRecipient(uint64 streamId) {
         require(
             msg.sender == streams[streamId].sender ||
                 msg.sender == streams[streamId].recipient,
@@ -15,7 +15,7 @@ contract Streaming {
         );
         _;
     }
-    modifier onlyValidateSreamId(uint256 streamId) {
+    modifier onlyValidateSreamId(uint64 streamId) {
         require(streamId <= streamIdCounter, "stream does not exist");
         _;
     }
@@ -32,7 +32,7 @@ contract Streaming {
     }
 
     event CreateStream(
-        uint256 indexed streamId,
+        uint64 indexed streamId,
         address indexed sender,
         address indexed recipient,
         uint256 deposit,
@@ -41,12 +41,12 @@ contract Streaming {
     );
 
     event WithdrawFromStream(
-        uint256 indexed streamId,
+        uint64 indexed streamId,
         address indexed recipient
     );
 
     event CancelStream(
-        uint256 indexed streamId,
+        uint64 indexed streamId,
         address indexed sender,
         address indexed recipient,
         uint256 vestedAmount,
@@ -62,7 +62,7 @@ contract Streaming {
         uint256 deposit,
         uint256 startTime,
         uint256 stopTime
-    ) external payable returns (uint256 streamId) {
+    ) external payable returns (uint64 streamId) {
         require(
             deposit == msg.value,
             "Please input the deposit equals your transfer amount"
@@ -87,7 +87,7 @@ contract Streaming {
         );
 
         streamIdCounter += 1;
-        uint256 currentStreamId = streamIdCounter;
+        uint64 currentStreamId = streamIdCounter;
 
         // Rate Per second
         uint256 rate = deposit / duration;
@@ -114,7 +114,7 @@ contract Streaming {
         return currentStreamId;
     }
 
-    function balanceOf(uint256 streamId, address who)
+    function balanceOf(uint64 streamId, address who)
         public
         view
         onlyValidateSreamId(streamId)
@@ -138,7 +138,7 @@ contract Streaming {
         }
     }
 
-    function elapsedTimeFor(uint256 streamId)
+    function elapsedTimeFor(uint64 streamId)
         private
         view
         onlyValidateSreamId(streamId)
@@ -157,7 +157,7 @@ contract Streaming {
         return stream.stopTime - stream.startTime;
     }
 
-    function withdrawFromStream(uint256 streamId)
+    function withdrawFromStream(uint64 streamId)
         external
         onlyValidateSreamId(streamId)
     {
@@ -182,7 +182,7 @@ contract Streaming {
         emit WithdrawFromStream(streamId, streams[streamId].recipient);
     }
 
-    function getStream(uint256 streamId)
+    function getStream(uint64 streamId)
         public
         view
         onlyValidateSreamId(streamId)
@@ -210,7 +210,7 @@ contract Streaming {
         );
     }
 
-    function cancelStream(uint256 streamId)
+    function cancelStream(uint64 streamId)
         external
         onlyValidateSreamId(streamId)
         onlySenderOrRecipient(streamId)
